Added ParagraphWrapper.php code in mid port state
Used js2php as provided by NPM to convert ParagraphWrapper.js
to php. The output has no indentation, lots of errors but
is still useful as a starting point. It looks better after
Phpstorm reformats it in a first pass. After that it needs lots
of manual format adjustment as well as lots of corrections.

Change-Id: I8a2ea5a5fd25db642980d11b2c2e3f9636fa7d5f